% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAIRADISE-package.R
\docType{package}
\name{PAIRADISE}
\alias{PAIRADISE}
\alias{PAIRADISE-package}
\title{PAIRADISE
Detecting allele-specific alternative splicing from population-scale RNA-seq data}
\description{
We introduce PAIRADISE (PAIred Replicate analysis of Allelic DIfferential Splicing Events), a method for detecting allele-specific alternative splicing (ASAS) from RNA-seq data. PAIRADISE uses a statistical model that aggregates ASAS signals across multiple individuals in a population. It formulates ASAS detection as a statistical problem for identifying differential alternative splicing from RNA-seq data with paired replicates. The PAIRADISE statistical model is applicable to many forms of allele-specific isoform variation (e.g. RNA editing), and can be used as a generic statistical model for RNA-seq studies involving paired replicates.
}
\seealso{
\code{\link{pairadise}}
}
