% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R, R/sample_dataset_CEU.R,
%   R/sample_dataset_LUSC.R
\docType{data}
\name{sample_dataset}
\alias{sample_dataset}
\alias{sample_dataset_CEU}
\alias{sample_dataset_LUSC}
\title{sample_dataset}
\format{
The dataset has 7 columns, arranged as follows:
\describe{
\item{ExonID}{Column 1 contains the ID of the alternative splicing events.}
\item{I1}{Column 2 contains counts of isoform 1 corresponding to the first group.}
\item{S1}{Column 3 contains counts of isoform 2 corresponding to the first group.}
\item{I2}{Column 4 contains counts of isoform 1 corresponding to the second group.}
\item{S2}{Column 5 contains counts of isoform 2 corresponding to the second group.}
\item{I_len}{Column 6 contains the effective length of isoform 1.}
\item{S_len}{Column 7 contains the effective length of isoform 2.}
}

The dataset has 7 columns, arranged as follows:
\describe{
\item{ExonID}{Column 1 contains the ID of the alternative splicing events.}
\item{I1}{Column 2 contains counts of isoform 1 corresponding to the first group.}
\item{S1}{Column 3 contains counts of isoform 2 corresponding to the first group.}
\item{I2}{Column 4 contains counts of isoform 1 corresponding to the second group.}
\item{S2}{Column 5 contains counts of isoform 2 corresponding to the second group.}
\item{I_len}{Column 6 contains the effective length of isoform 1.}
\item{S_len}{Column 7 contains the effective length of isoform 2.}
}

The dataset has 7 columns, arranged as follows:
\describe{
\item{ExonID}{Column 1 contains the ID of the alternative splicing events.}
\item{I1}{Column 2 contains counts of isoform 1 corresponding to the first group.}
\item{S1}{Column 3 contains counts of isoform 2 corresponding to the first group.}
\item{I2}{Column 4 contains counts of isoform 1 corresponding to the second group.}
\item{S2}{Column 5 contains counts of isoform 2 corresponding to the second group.}
\item{I_len}{Column 6 contains the effective length of isoform 1.}
\item{S_len}{Column 7 contains the effective length of isoform 2.}
}
}
\usage{
data(sample_dataset)

data(sample_dataset_CEU)

data(sample_dataset_LUSC)
}
\description{
The CEU dataset was generated by analyzing the allele-specific
alternative splicing events in the GEUVADIS CEU
data. Allele-specific reads were mapped onto alternative splicing
events using rPGA (version 2.0.0). Then the allele-specific bam
files mapped onto the two haplotypes are merged together to detect
alternative splicing events using rMATS (version 3.2.5)16.

The LUSC dataset was generated by analyzing the tumor versus
adjacent control samples from TCGA LUSC RNA-seq data.
}
