% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R, R/sample_dataset_CEU.R,
%   R/sample_dataset_LUSC.R
\docType{data}
\name{sample_dataset}
\alias{sample_dataset}
\alias{sample_dataset_CEU}
\alias{sample_dataset_LUSC}
\title{sample_dataset}
\format{The dataset has 7 columns, arranged as follows:
\describe{
\item{ExonID}{Column 1 contains the ID of the alternative splicing events.}
\item{I1}{Column 2 contains counts of isoform 1 corresponding to the first group.}
\item{S1}{Column 3 contains counts of isoform 2 corresponding to the first group.}
\item{I2}{Column 4 contains counts of isoform 1 corresponding to the second group.}
\item{S2}{Column 5 contains counts of isoform 2 corresponding to the second group.}
\item{I_len}{Column 6 contains the effective length of isoform 1.}
\item{S_len}{Column 7 contains the effective length of isoform 2.}
}}
\usage{
data(sample_dataset)

data(sample_dataset_CEU)

data(sample_dataset_LUSC)
}
\description{
sample_dataset

sample_dataset_CEU

sample_dataset_LUSC
}
